requirements.txt
fyers-apiv3==3.0.9
py5paisa==0.7.12
pyotp==2.9.0
setuptools==69.2.0
python-dotenv
python-telegram-bot==13.7
flask
flask-login
